MB-61093 Add method to compute distance from codes for IVF index
Given a query vector x, compute distance to provided codes
for the input list_no. This is a special purpose method
to be used as a flat distance computer for an inverted
list where codes are provided externally. This allows to
use the quantizer while computing distance for the quantized
codes.