MB-61093 Add method to compute distance from codes for IVF index
* Added new function compute_distance_to_codes_for_list to IndexIVF.
Given a query vector x, this function computes distance to provided
codes for the input list_no. This is a special purpose method
to be used as a flat distance computer for an inverted
list where codes are provided externally. This allows to
use the quantizer independently while computing distance for
the quantized codes.

* This function is currently implemented only for IndexIVFScalarQuantizer.

* Extended c_api interface for the new function.